+++ tomcat/src/main/java/org/apache/coyote/http11/request/line/HttpMethod.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.request;
+package org.apache.coyote.http11.request.line;
 
 public enum HttpMethod {
     GET, POST, DELETE, PATCH, PUT
