+++ tomcat/src/main/java/org/apache/coyote/http11/response/header/HttpHeader.java
@@ -1,4 +1,4 @@
-package org.apache.coyote.http11.response;
+package org.apache.coyote.http11.response.header;
 
 public enum HttpHeader {
     CONTENT_TYPE("Content-Type"),
