+++ tomcat/src/main/java/org/apache/coyote/http11/response/line/ResponseStatus.java
@@ -1,10 +1,8 @@
-package org.apache.coyote.http11.response;
+package org.apache.coyote.http11.response.line;
 
 public enum ResponseStatus {
     OK(200, "OK"),
     FOUND(302, "Found"),
-    BAD_REQUEST(400, "Bad Request"),
-
     UNAUTHORIZED(401, "Unauthorized"),
     ;
