+++ tomcat/src/test/java/org/apache/coyote/http11/HttpCookieTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.SoftAssertions.assertSoftly;
 
 import java.util.Map;
-import org.apache.coyote.http11.request.HttpCookie;
+import org.apache.coyote.http11.request.header.HttpCookie;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
