+++ tomcat/src/test/java/org/apache/coyote/http11/request/HttpRequestTest.java
@@ -1,8 +1,10 @@
 package org.apache.coyote.http11.request;
 
-import static org.apache.coyote.http11.request.HttpMethod.POST;
+import static org.apache.coyote.http11.request.line.HttpMethod.POST;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.apache.coyote.http11.request.header.RequestHeader;
+import org.apache.coyote.http11.request.line.RequestLine;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Nested;
