+++ tomcat/src/test/java/org/apache/coyote/http11/request/RequestBodyTest.java
@@ -3,6 +3,7 @@ package org.apache.coyote.http11.request;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import java.util.Map;
+import org.apache.coyote.http11.request.body.RequestBody;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
