+++ tomcat/src/test/java/org/apache/coyote/http11/request/RequestLineTest.java
@@ -1,8 +1,10 @@
 package org.apache.coyote.http11.request;
 
-import static org.apache.coyote.http11.request.HttpMethod.GET;
+import static org.apache.coyote.http11.request.line.HttpMethod.GET;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.apache.coyote.http11.request.line.RequestLine;
+import org.apache.coyote.http11.request.line.RequestUri;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
