+++ tomcat/src/test/java/org/apache/coyote/http11/request/RequestUriTest.java
@@ -2,6 +2,7 @@ package org.apache.coyote.http11.request;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.apache.coyote.http11.request.line.RequestUri;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
