+++ tomcat/src/test/java/org/apache/coyote/http11/response/HttpContentTypeTest.java
@@ -3,6 +3,7 @@ package org.apache.coyote.http11.response;
 import static java.nio.charset.StandardCharsets.UTF_8;
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.apache.coyote.http11.response.header.HttpContentType;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
