+++ tomcat/src/test/java/org/apache/coyote/http11/response/ResponseBodyTest.java
@@ -1,10 +1,11 @@
 package org.apache.coyote.http11.response;
 
+import static org.assertj.core.api.Assertions.assertThat;
+
+import org.apache.coyote.http11.response.body.ResponseBody;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
-
-import static org.assertj.core.api.Assertions.assertThat;
 
 @SuppressWarnings("NonAsciiCharacters")
 @DisplayNameGeneration(DisplayNameGenerator.ReplaceUnderscores.class)
